Fix Pages Content page

+++ web-interface/src/views/content/languages/Languages.js
@@ -68,6 +68,7 @@ const Languages = () => {
   }
 
   const handleRemove = async (id) => {
+    if (!window.confirm('Are you sure you want to delete this language?')) return
     await deleteLanguage(id)
     fetchAndSetLanguages()
   }
